Move UserMenu outside of Navigation
