+++ src/components/Navigation/index.js
@@ -1,27 +1,13 @@
 import PropTypes from "prop-types";
-import React, { useState } from "react";
-
-import { Menu } from "../Menu";
+import React from "react";
 
 import { AppSwitcher } from "./appSwitcher";
+import { UserMenu } from "./userMenu";
 import { NavigationItem } from "./navigationItem";
 
 import "./styles.css";
 
 export const Navigation = ({ appswitcherItems, navigationItems, supportWebsiteURL, usermenuAvatar, usermenuItems }) => {
-
-  // Usermenu state
-  const [isOpenUsermenu, setOpenUsermenu] = useState(false)
-
-  const closeUsermenu = () => {
-    document.removeEventListener("click", closeUsermenu, false);
-    setOpenUsermenu(false);
-  };
-
-  const openUsermenu = () => {
-    document.addEventListener("click", closeUsermenu, false);
-    setOpenUsermenu(true);
-  };
 
   return (
     <nav className="navigation">
@@ -46,18 +32,7 @@ export const Navigation = ({ appswitcherItems, navigationItems, supportWebsiteUR
       </div>
       <div>
         <NavigationItem icon="H" isExternal link={supportWebsiteURL} title="Support Website" />
-        <div style={{ position: "relative" }}>
-          <button
-            className="navigation-usermenu tooltip-anchor"
-            onClick={() => { isOpenUsermenu ? closeUsermenu() : openUsermenu(); }}>
-            {usermenuAvatar}
-          </button>
-          <span className="navigation-usermenu-caret">></span>
-          <div className="tooltip">User Settings</div>
-          {isOpenUsermenu && (
-            <Menu menuItems={usermenuItems} />
-          )}
-        </div>
+        <UserMenu usermenuItems={usermenuItems} usermenuAvatar={usermenuAvatar} />
       </div>
     </nav>
   );
